refactor: change file from cjs to esm
